+++ backend/app/auth/db/password/model.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import ForeignKey, LargeBinary, String
+from sqlalchemy import ForeignKey, LargeBinary
 from sqlalchemy.orm import Mapped, mapped_column
 from sqlalchemy.sql.functions import now
 
@@ -14,13 +14,4 @@ class Password(Base):
     id: Mapped[int] = mapped_column(primary_key=True, autoincrement=True)
     user_id: Mapped[int] = mapped_column(ForeignKey("users.id", ondelete=CASCADE), nullable=False)
     value: Mapped[bytes] = mapped_column(LargeBinary(100), nullable=False)
-    created_at: Mapped[datetime] = mapped_column(server_default=now(), nullable=False)
-
-
-class IssuedToken(Base):
-    __tablename__ = "issued_tokens"
-
-    value: Mapped[str] = mapped_column(String(512), primary_key=True)
-    user_id: Mapped[int] = mapped_column(ForeignKey("users.id", ondelete=CASCADE), nullable=False)
-    revoked: Mapped[bool] = mapped_column(default=False, nullable=False)
     created_at: Mapped[datetime] = mapped_column(server_default=now(), nullable=False)
