+++ backend/app/meal_requests/db/meal_request/model.py
@@ -15,14 +15,3 @@ class MealRequest(Base):
     teacher_id: Mapped[int] = mapped_column(ForeignKey("users.id", ondelete=CASCADE), nullable=False)
     meal_id: Mapped[int] = mapped_column(ForeignKey("meals.id"), unique=True, nullable=False)
     created_at: Mapped[datetime] = mapped_column(server_default=now(), nullable=False)
-
-
-class DeclaredPupils(Base):
-    __tablename__ = "declared_pupils"
-
-    request_id: Mapped[int] = mapped_column(ForeignKey("meal_requests.id", ondelete=CASCADE), primary_key=True)
-    pupil_id: Mapped[str] = mapped_column(ForeignKey("pupils.id", ondelete=CASCADE), primary_key=True)
-    breakfast: Mapped[bool] = mapped_column(nullable=False)
-    lunch: Mapped[bool] = mapped_column(nullable=False)
-    dinner: Mapped[bool] = mapped_column(nullable=False)
-    preferential: Mapped[bool] = mapped_column(nullable=False)
