+++ backend/app/pupils/domain/base_repositories.py
@@ -1,7 +1,7 @@
 from abc import ABC
 
 from app.database.base import Repository
-from app.pupils.db.models import Pupil
+from app.pupils.db.pupil.model import Pupil
 
 
 class BasePupilsRepository(Repository[Pupil], ABC):
