+++ backend/app/users/domain/base_repositories.py
@@ -1,7 +1,7 @@
 from abc import ABC
 
 from app.database.base import Repository
-from app.users.db.models import User
+from app.users.db.user.model import User
 
 
 class BaseUsersRepository(Repository[User], ABC):
