+++ backend/tests/auth/integration/conftest.py
@@ -8,9 +8,10 @@ from httpx import Response
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.sql.functions import now
 
-from app.auth.db.models import IssuedToken, Password
+from app.auth.db.issued_token.model import IssuedToken
+from app.auth.db.password.model import Password
 from app.config import JWTSettings
-from app.users.db.models import Role, User
+from app.users.db.user.model import Role, User
 from tests.utils import get_set_cookies
